hardware/OLD_Project/DEPRECATED/requirements.txt
aioice==0.9.0
aiortc==1.5.0
alembic==1.12.0
av==10.0.0
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.4
dnspython==2.4.2
eventlet==0.33.3
Flask==3.0.0
Flask-Cors==4.0.0
flask-crontab==0.1.2
Flask-Migrate==4.0.5
Flask-Script==2.0.6
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
google-crc32c==1.5.0
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
ifaddr==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
ndg-httpsclient==0.5.1
numpy==1.26.0
opencv-python==4.8.1.78
packaging==23.2
pusher==3.3.2
pyasn1==0.5.0
pycparser==2.21
pyee==11.0.0
pylibsrtp==0.8.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
python-engineio==4.7.1
python-socketio==5.9.0
requests==2.31.0
simple-websocket==0.10.1
six==1.16.0
SQLAlchemy==2.0.21
SQLAlchemy-Utils==0.41.1
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==3.0.0
wsproto==1.2.0
